backend/requirements.txt
torch
torchvision
torchaudio
transformers
datasets
pillow
numpy==1.26.4
scikit-image
imageio
imageio-ffmpeg
moviepy
fastapi
uvicorn
python-multipart
librosa
soundfile
mediapipe
tqdm
matplotlib
